Bug 970271: Fix GCC warnings about comparison of signed and unsigned values, r=rjesup

+++ netwerk/srtp/src/crypto/replay/rdb.c
@@ -104,7 +104,7 @@ rdb_add_index(rdb_t *rdb, uint32_t p_index) {
   /* here we *assume* that p_index > rdb->window_start */
 
   delta = (p_index - rdb->window_start);    
-  if (delta < rdb_bits_in_bitmask) {
+  if (delta < (int)rdb_bits_in_bitmask) {
 
     /* if the p_index is within the window, set the appropriate bit */
     v128_set_bit(&rdb->bitmask, delta);
